+++ streamingpro-mlsql/src/main/java/streaming/dsl/auth/CreateAuth.scala
@@ -23,7 +23,7 @@ import org.apache.spark.sql.execution.MLSQLAuthParser
 import streaming.dsl.parser.DSLSQLLexer
 import streaming.dsl.parser.DSLSQLParser._
 import streaming.dsl.template.TemplateMerge
-import streaming.dsl.DslTool
+import tech.mlsql.dsl.adaptor.DslTool
 import tech.mlsql.dsl.processor.AuthProcessListener
