+++ streamingpro-mlsql/src/main/java/streaming/dsl/auth/SelectAuth.scala
@@ -20,10 +20,10 @@ package streaming.dsl.auth
 
 import org.antlr.v4.runtime.misc.Interval
 import org.apache.spark.sql.execution.MLSQLAuthParser
-import streaming.dsl.DslTool
 import streaming.dsl.parser.DSLSQLLexer
 import streaming.dsl.parser.DSLSQLParser._
 import streaming.dsl.template.TemplateMerge
+import tech.mlsql.dsl.adaptor.DslTool
 import tech.mlsql.dsl.processor.AuthProcessListener
