+++ streamingpro-mlsql/src/main/java/tech/mlsql/dsl/adaptor/DropAdaptor.scala
@@ -16,9 +16,10 @@
  * limitations under the License.
  */
 
-package streaming.dsl
+package tech.mlsql.dsl.adaptor
 
 import org.antlr.v4.runtime.misc.Interval
+import streaming.dsl.ScriptSQLExecListener
 import streaming.dsl.parser.DSLSQLLexer
 import streaming.dsl.parser.DSLSQLParser.SqlContext
 import streaming.dsl.template.TemplateMerge
