+++ streamingpro-mlsql/src/main/java/tech/mlsql/dsl/adaptor/IncludeAdaptor.scala
@@ -16,11 +16,11 @@
  * limitations under the License.
  */
 
-package streaming.dsl
+package tech.mlsql.dsl.adaptor
 
+import streaming.dsl.IncludeSource
 import streaming.dsl.parser.DSLSQLParser._
 import streaming.dsl.template.TemplateMerge
-import tech.mlsql.dsl.adaptor.{PreProcessIncludeListener, SCType}
 
 /**
   * Created by allwefantasy on 12/1/2018.
