+++ streamingpro-mlsql/src/main/java/tech/mlsql/dsl/adaptor/RegisterAdaptor.scala
@@ -16,10 +16,11 @@
  * limitations under the License.
  */
 
-package streaming.dsl
+package tech.mlsql.dsl.adaptor
 
 import java.util.UUID
 
+import streaming.dsl.ScriptSQLExecListener
 import streaming.dsl.parser.DSLSQLParser._
 import streaming.dsl.template.TemplateMerge
