+++ streamingpro-mlsql/src/main/java/tech/mlsql/dsl/adaptor/SetAdaptor.scala
@@ -16,10 +16,11 @@
  * limitations under the License.
  */
 
-package streaming.dsl
+package tech.mlsql.dsl.adaptor
 
 import _root_.streaming.dsl.parser.DSLSQLParser._
 import streaming.common.ShellCommand
+import streaming.dsl.ScriptSQLExecListener
 import streaming.dsl.template.TemplateMerge
 import tech.mlsql.Stage
